+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This change
 - bugfix in handling options in notebook generation
 - bugfix: handling the edge case of string hiccup carefully
 - bugfix: handling `:kind/hidden` correctly in notebook generation
+- applying `:kindly/options` in all hiccup-based pathways
 
 ## [2-beta15] - 2024-07-27
 - introducing code-and-value and horizontal layout - WIP (PR #127)